Release intermediate *v8.Value instances during v8.Create.
This relieves pressure on the GC since it doesn't have to track these
finalizers anymore and we know that there no more references to them.
This is an optmization: if v.release() is not called due to an error,
it will eventually be called by the GC.

Preliminary tests of an upcoming API to query the v8 memory state
indicate that this improves the v8.Create performance.

See #17 for more information.