Fatten thin archives when creating docker libs.
Some versions of V8 appear to produce thin archives during compilation
instead of the expected fat archives. That means they don't contain
all of the code but instead reference other object files on disk.
That's annoying since it makes it difficult to know what to include in
the final lib directory.

This adds a script that attempts to fattenify all of the produced
archives. It will ignore any failures if the archives are already fat.

References:
- https://github.com/fluxio/go-v8
- https://stackoverflow.com/questions/25554621/turn-thin-archive-into-normal-one